[NET-526] Nested Toast component in Toast container, added default config
