+++ app/src/app/layout/layout.module.ts
@@ -66,7 +66,6 @@ import { GeneralModalComponent } from './helpers/modals/general-modal/general-mo
 import { RenderModule } from '../framework/render/render.module';
 import { ObjectPreviewModalComponent} from '../framework/object/modals/object-preview-modal/object-preview-modal.component';
 import { InfoBoxComponent } from './components/info-box/info-box.component';
-import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
 
 export const LAYOUT_COMPONENT_ROUTES: Routes = [
   {
