+++ app/src/app/layout/toast/toast-container/toast-component/toast-component.component.html
@@ -1,0 +1,19 @@
+<div class="toast-container shadow mb-2 {{toast.classname}}">
+  <span>{{waitTillDisposal((toast.delayTime * 1000) || 5000, toast)}}</span>
+  <div class="toast-body border-0">
+    <div class="container">
+      <div class="row">
+        <div class="col-2 pl-0"><i class="{{toast?.icon}} {{toast.headerClass}} fa-3x fa-align-center icon-center"></i></div>
+        <div class="col-9"><span><h5>{{toast?.headerName}}</h5> {{toast?.text}}</span></div>
+        <div class="col-1 pr-0">
+          <button type="button" class="ml-2 mb-2 close" data-dismiss="toast" aria-label="Close">
+            <span aria-hidden="true" (click)="toastService?.remove(toast)">×</span>
+          </button>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="meter">
+    <span style="width: 100%"><span class="progress" [style.animation]="parse(toast?.delayTime)"></span></span>
+  </div>
+</div>
