+++ app/src/app/layout/toast/toast-container/toast-container.component.html
@@ -1,119 +1,31 @@
-<div class="right fixed">
+<div class="right fixed ">
   <div *ngFor="let toast of toastService.toastsright"
-       class="toast  bg-white {{toast.classname}} {{toast.direction}}"
        [@simpleFadeAnimation]="'in'">
-    <span>{{waitTillDisposal((toast.delayTime * 1000) || 5000, toast)}}</span>
-    <div class="toast-body ">
-      <div class="container">
-        <div class="row">
-          <div class="col-2 pl-0"><i class="{{toast.icon}} {{toast.headerClass}} fa-3x fa-align-center icon-center"></i></div>
-          <div class="col-9"><span><h5>{{toast.headerName}}</h5> {{toast.text}}</span></div>
-          <div class="col-1 pr-0">
-            <button type="button" class="ml-2 mb-2 close" data-dismiss="toast" aria-label="Close">
-              <span aria-hidden="true" (click)="toastService.remove(toast)">×</span>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="meter">
-      <span style="width: 100%"><span class="progress" [style.animation]="parse(toast.delayTime)"></span></span>
-    </div>
+    <cmdb-toast-component [toast]="toast"></cmdb-toast-component>
   </div>
 </div>
 
 <div class="left fixed">
-  <div *ngFor="let toast of toastService.toastsleft"
-       class="toast  bg-white {{toast.classname}} {{toast.direction}}"
-       [@simpleFadeAnimation]="'in'">
-    <span>{{waitTillDisposal((toast.delayTime * 1000) || 5000, toast)}}</span>
-    <div class="toast-body ">
-      <div class="container">
-        <div class="row">
-          <div class="col-2 pl-0"><i class="{{toast.icon}} {{toast.headerClass}} fa-3x fa-align-center icon-center"></i></div>
-          <div class="col-9"><span><h5>{{toast.headerName}}</h5> {{toast.text}}</span></div>
-          <div class="col-1 pr-0">
-            <button type="button" class="ml-2 mb-2 close" data-dismiss="toast" aria-label="Close">
-              <span aria-hidden="true" (click)="toastService.remove(toast)">×</span>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="meter">
-      <span style="width: 100%"><span class="progress" [style.animation]="parse(toast.delayTime)"></span></span>
-    </div>
+  <div *ngFor="let toast of toastService.toastsleft" [@simpleFadeAnimation]="'in'">
+    <cmdb-toast-component [toast]="toast"></cmdb-toast-component>
   </div>
 </div>
 
 <div class="bottom right fixed">
-  <div *ngFor="let toast of toastService.toastsdownright"
-       class="toast  bg-white {{toast.classname}} {{toast.direction}}"
-       [@simpleFadeAnimation]="'in'">
-    <span>{{waitTillDisposal((toast.delayTime * 1000) || 5000, toast)}}</span>
-    <div class="toast-body ">
-      <div class="container">
-        <div class="row">
-          <div class="col-2 pl-0"><i class="{{toast.icon}} {{toast.headerClass}} fa-3x fa-align-center icon-center"></i></div>
-          <div class="col-9"><span><h5>{{toast.headerName}}</h5> {{toast.text}}</span></div>
-          <div class="col-1 pr-0">
-            <button type="button" class="ml-2 mb-2 close" data-dismiss="toast" aria-label="Close">
-              <span aria-hidden="true" (click)="toastService.remove(toast)">×</span>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="meter">
-      <span style="width: 100%"><span class="progress" [style.animation]="parse(toast.delayTime)"></span></span>
-    </div>
+  <div *ngFor="let toast of toastService.toastsdownright" [@simpleFadeAnimation]="'in'">
+    <cmdb-toast-component [toast]="toast"></cmdb-toast-component>
   </div>
 </div>
 
 <div class="bottom left fixed">
-  <div *ngFor="let toast of toastService.toastsdownleft"
-       class="toast  bg-white {{toast.classname}} {{toast.direction}}"
-       [@simpleFadeAnimation]="'in'">
-    <span>{{waitTillDisposal((toast.delayTime * 1000) || 5000, toast)}}</span>
-    <div class="toast-body ">
-      <div class="container">
-        <div class="row">
-          <div class="col-2 pl-0"><i class="{{toast.icon}} {{toast.headerClass}} fa-3x fa-align-center icon-center"></i></div>
-          <div class="col-9"><span><h5>{{toast.headerName}}</h5> {{toast.text}}</span></div>
-          <div class="col-1 pr-0">
-            <button type="button" class="ml-2 mb-2 close" data-dismiss="toast" aria-label="Close">
-              <span aria-hidden="true" (click)="toastService.remove(toast)">×</span>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="meter">
-      <span style="width: 100%"><span class="progress" [style.animation]="parse(toast.delayTime)"></span></span>
-    </div>
+  <div *ngFor="let toast of toastService.toastsdownleft" [@simpleFadeAnimation]="'in'">
+    <cmdb-toast-component [toast]="toast"></cmdb-toast-component>
   </div>
 </div>
 
 <div class="center fixed">
-  <div *ngFor="let toast of toastService.toastscenter"
-       class="toast  bg-white {{toast.classname}} {{toast.direction}}"
-       [@simpleFadeAnimation]="'in'">
-    <span>{{waitTillDisposal((toast.delayTime * 1000) || 5000, toast)}}</span>
-    <div class="toast-body ">
-      <div class="container">
-        <div class="row">
-          <div class="col-2 pl-0"><i class="{{toast.icon}} {{toast.headerClass}} fa-3x fa-align-center icon-center"></i></div>
-          <div class="col-9"><span><h5>{{toast.headerName}}</h5> {{toast.text}}</span></div>
-          <div class="col-1 pr-0">
-            <button type="button" class="ml-2 mb-2 close" data-dismiss="toast" aria-label="Close">
-              <span aria-hidden="true" (click)="toastService.remove(toast)">×</span>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="meter">
-      <span style="width: 100%"><span class="progress" [style.animation]="parse(toast.delayTime)"></span></span>
-    </div>
+  <div *ngFor="let toast of toastService.toastscenter" [@simpleFadeAnimation]="'in'">
+    <cmdb-toast-component [toast]="toast"></cmdb-toast-component>
   </div>
 </div>
+
